export data sources in datapackage
