+++ intermine/webapp/src/main/webapp/WEB-INF/global.web.properties
@@ -69,6 +69,7 @@ begin.thirdBox.linkTitle = Tell me more!
 #### HEAD - configuration for the head element of pages, mainly scripts and style-sheets ###
 
 head.cdn.location = https://cdn.intermine.org
+head.imTables.location = https://0780b2d9f9ab.ngrok.io
 
 head.css.all.A = inlineTagEditor.css
 head.css.all.B = resultstables.css
@@ -116,8 +117,8 @@ head.js.results.A = jquery.qtip-1.0.0-rc3.min.js
 head.js.results.B = raphael.js
 head.js.results.C = jsphylosvg.js
 head.js.results.H = list-listener.js
-head.js.results.RESULT_TABLES = CDN/js/intermine/im-tables/2.1.x/imtables.js
-head.css.results.RESULT_TABLES = CDN/js/intermine/im-tables/2.1.x/main.sandboxed.css
+head.js.results.RESULT_TABLES = imTables/dist/imtables.js
+head.css.results.RESULT_TABLES = imTables/dist/main.sandboxed.css
 
 head.js.begin.A = jQuery.roundCorners-1.1.1.js
 head.js.begin.B = excanvas.js
